Purchase Order Extension
Notes: Docx xml doesnt work like a variable instead when a create a layoyt from AL this make a new DATA SET. In this problem i modify the exising purchase line to add a tirgger that run every time a new row is added.
As for the Total amount i added it to the Total amount which is another dataaset.

The reason is on DOCX xml port, we get the last row. So if we calculate the total amount before hand set it to the TOTAL this works just fine.